Add the "plot" module
