+++ src/atldld/cli.py
@@ -128,9 +128,8 @@ def dataset_info(dataset_id):
 @click.argument("dataset_id", type=int)
 def dataset_preview(dataset_id):
     import numpy as np
-    from matplotlib.figure import Figure
 
-    from atldld import constants, requests
+    from atldld import plot, requests
 
     # Send request
     rma_parameters = requests.RMAParameters(
@@ -156,8 +155,8 @@ def dataset_preview(dataset_id):
 
     meta = msg[0]
     section_image_metas = meta.pop("section_images")
-    section_image_metas.sort(key=lambda image_meta: image_meta["section_number"])
-    section_image_metas = section_image_metas[:3]
+    section_image_metas.sort(key=lambda image_meta_: image_meta_["section_number"])
+    section_image_metas = section_image_metas  # [:2]
 
     click.secho("Fetching the corner coordinates of the section images...", fg="green")
     all_corners = []
@@ -170,59 +169,9 @@ def dataset_preview(dataset_id):
             )
             all_corners.append(np.array(corners))
 
-    def draw_slice_2d(ax, points):
-        coords = points.T
-        ax.plot(*coords, color="green")
-        ax.scatter(*coords, color="red")
-
     click.secho("Plotting...", fg="green")
-    scale = 25
-    n_p, n_i, n_r = np.array(constants.REF_DIM_1UM) / scale
-
-    fig = Figure(figsize=(14, 4), dpi=200)
-    axs = fig.subplots(
-        ncols=4,
-        sharey=True,
-        gridspec_kw={"width_ratios": [16 / 7, 1, 16 / 7, 1]}
-    )
+    fig = plot.preview_sagittal_dataset(all_corners)
     fig.suptitle(f"Dataset ID {dataset_id}", fontsize=32)
-    for ax in axs.ravel():
-        ax.grid(True, linestyle=":", color="gray")
-        ax.set_ylim((0, n_r))
-    #     ax.set_aspect("equal")
-    ax1, ax2, ax3, ax4 = axs.ravel()
-
-    ax1.set_title("Edge $(0, 0)-(1, 0)$")
-    ax1.set_xlabel("p (coronal)", fontsize=16)
-    ax1.set_ylabel("r (sagittal)", fontsize=16)
-    ax1.axvline(0, color="blue", linestyle=":")
-    ax1.axvline(n_p, color="blue", linestyle=":")
-    for corners in all_corners:
-        draw_slice_2d(ax1, corners[np.ix_([0, 1], [0, 2])] / scale)
-
-    ax2.set_title("Edge $(1, 0)-(1, 1)$")
-    ax2.set_xlabel("i (transversal)", fontsize=16)
-    ax2.axvline(0, color="blue", linestyle=":")
-    ax2.axvline(n_i, color="blue", linestyle=":")
-    for corners in all_corners:
-        draw_slice_2d(ax2, corners[np.ix_([1, 2], [1, 2])] / scale)
-
-    ax3.set_title("Edge $(1, 1)-(0, 1)$")
-    ax3.set_xlabel("p (coronal)", fontsize=16)
-    for corners in all_corners:
-        draw_slice_2d(ax3, corners[np.ix_([2, 3], [0, 2])] / scale)
-    ax3.axvline(0, color="blue", linestyle=":")
-    ax3.axvline(n_p, color="blue", linestyle=":")
-    ax3.invert_xaxis()
-
-    ax4.set_title("Edge $(0, 1)-(0, 0)$")
-    ax4.set_xlabel("i (transversal)", fontsize=16)
-    for corners in all_corners:
-        draw_slice_2d(ax4, corners[np.ix_([3, 0], [1, 2])] / scale)
-    ax4.axvline(0, color="blue", linestyle=":")
-    ax4.axvline(n_i, color="blue", linestyle=":")
-    ax4.invert_xaxis()
-    fig.tight_layout()
     fig.savefig(f"dataset-id-{dataset_id}-preview.png")
